refactor
split main into separate modules to improve readability, added download stats
